добавлен README.md

+++ src/main/java/com/tasks/tasks/restController/TasksController.java
@@ -1,15 +1,13 @@
 package com.tasks.tasks.restController;
 
-import java.sql.Time;
-import java.sql.Timestamp;
-import java.time.Instant;
-import java.util.Date;
-import java.util.Optional;
-import com.tasks.tasks.core.TaskRepository;
 import com.tasks.tasks.core.TaskItem;
+import com.tasks.tasks.core.TaskRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
+
+import java.time.Instant;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/v2")
@@ -28,32 +26,40 @@ public class TasksController {
 
     @GetMapping("/list/{id}")
     public Optional<TaskItem> getTaskById ( @PathVariable("id") Integer id) {
-        return this.taskRepository.findById(id);
+
+        if (this.taskRepository.findById(id).isPresent()) {
+            return this.taskRepository.findById(id);
+        }
+
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
     }
 
     @DeleteMapping("/del/{id}")
     public void deleteTaskById (@PathVariable("id") Integer id) {
 
-        try {
+        if (this.taskRepository.findById(id).isPresent()) {
             this.taskRepository.deleteById(id);
+            throw new ResponseStatusException(HttpStatus.OK);
         }
-        catch(Exception e) {
-            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR);
+        else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
 
-        throw new ResponseStatusException(HttpStatus.OK);
 
     }
 
     @PostMapping("/new")
     public void addNewTask(@RequestBody TaskItem testTaskItem) {
-        //this.taskRepository.save(testTaskItem);
+
+        testTaskItem.setCreateTime(Instant.now().getEpochSecond());
         TaskItem newTestTaskItem = this.taskRepository.save(testTaskItem);
 
-        long unixTime = Instant.now().getEpochSecond();
-        System.out.println(unixTime);
+        if (this.taskRepository.findById(testTaskItem.getId()).isPresent()) {
+            throw new ResponseStatusException(HttpStatus.CREATED);
+        }
 
-        throw new ResponseStatusException(HttpStatus.CREATED);
+        throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR);
     }
 
     @PutMapping("/mod/{id}")
@@ -62,7 +68,7 @@ public class TasksController {
 
         Optional<TaskItem> taskToUpdateOptional = this.taskRepository.findById(id);
         if (!taskToUpdateOptional.isPresent()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Task not found");
             //return null;
         }
         TaskItem taskToUpdate = taskToUpdateOptional.get();
